+++ src/main/java/duke.txt
@@ -1,1 +1,3 @@
-E|0|CS2113 lecture |Friday 18 September 2020 from 4pm to 6pm
+T|1|buy bread
+T|0|meeting
+T|0|project outing
