+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,4 @@
-package duke.command;
+package duke;
 
 public class DukeException extends Exception {
     protected String error;
